AuthContext is separated from AuthProvider.tsx AuthContext null checking is added on each files
