+++ src/__tests__/index.test.tsx
@@ -14,20 +14,10 @@
  * limitations under the License.
  */
 
-import {AuthContextInterface} from '../types';
+describe('PrivateRoute', () => {
+  it('is truthy', () => {
 
-/**
- * Utility Function
- * @description Used to check if the auth provider is missing or not
- * @throws Error if the auth provider is missing
- * @param authContext
- */
-function checkAuthProvider(authContext: AuthContextInterface):void {
-  if (authContext === null || authContext === undefined) {
-    throw new
-    Error('Auth Provider is missing. ' +
-      'Please add the AuthProvider before Router');
-  }
-}
+  });
+});
 
-export default checkAuthProvider;
+export {};
