+++ src/higherOrderComponents/withSignOut.tsx
@@ -5,8 +5,7 @@
   *@license Apache-2.0
   */
 import * as React from 'react';
-import {AuthContextConsumer} from '../AuthProvider';
-import checkAuthProvider from '../utils/checkAuthProvider';
+import {AuthContextConsumer} from '../AuthContext';
 
 /**
  * @interface withSignOutProps
@@ -29,7 +28,11 @@ function withSignOut<P extends withSignOutProps>(
     return (
       <AuthContextConsumer>
         {(c) => {
-          checkAuthProvider(c);
+          if (c === null) {
+            throw new
+            Error('Auth Provider is missing. ' +
+              'Please add the AuthProvider before Router');
+          }
           const signOut = () => {
             try {
               if (c?.authState.authToken) {
