+++ src/hooks/useAuth.ts
@@ -7,7 +7,7 @@
 
 
 import * as React from 'react';
-import {AuthContext} from '../AuthProvider';
+import AuthContext from '../AuthContext';
 import {signInFunctionParams} from '../types';
 
 /**
@@ -27,6 +27,11 @@ function useAuth():
     *@kind constant
     */
   const c = React.useContext(AuthContext);
+  if (c === null) {
+    throw new
+    Error('Auth Provider is missing. ' +
+      'Please add the AuthProvider before Router');
+  }
 
   /** @function
     * @name authHeader
