--- src/index.test.tsx
@@ -1,7 +1,0 @@
-import PrivateRoute from './PrivateRoute';
-
-describe('PrivateRoute', () => {
-  it('is truthy', () => {
-    expect(PrivateRoute).toBeTruthy();
-  });
-});
